dataproc_batch_trigger_fn/requirements.txt
google-cloud-dataproc==5.22.0
